pyproject.toml
[project]
name = "petcam-lab"
version = "0.1.0"
description = "도마뱀 특화 펫캠(게코 캠) 영상 백엔드 서비스"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.43.2",
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "opencv-python>=4.13.0.92",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "moto[s3]>=5.1.22",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
# 프로젝트 루트를 sys.path에 추가 → `from backend.capture import ...` 가능해짐
pythonpath = ["."]
testpaths = ["tests"]
